+++ src/common/classes/ah-class.class.js
@@ -376,6 +376,7 @@ export class ahClass {
         this.#args = args;
         this.#opts = new ahOptions( args );
         //logger.debug( 'instanciating', this.name(), args );
+        //console.trace( 'instanciating', this.name(), args );
 
         // if the name is already instanciated, then just return it
         const instance = AccountsHub.getInstance( this.name());
@@ -396,9 +397,6 @@ export class ahClass {
         }
         // and (on the server) deny all client-side direct updates
         this._deny();
-
-        // timestamp to verify the received data on HMR
-        this._stamp = Date.now();
 
         // register this new instance
         AccountsHub.getInstance( this.name(), this );