+++ src/common/js/configure.js
@@ -6,6 +6,7 @@ import _ from 'lodash';
 
 import { Logger } from 'meteor/pwix:logger';
 import { ReactiveVar } from 'meteor/reactive-var';
+import { Tracker } from 'meteor/tracker';
 
 const logger = Logger.get();
 
@@ -46,3 +47,17 @@ AccountsHub.configure = function( o ){
 
 _conf = _.merge( {}, AccountsHub._defaults );
 AccountsHub._conf.set( _conf );
+
+// we warn when 'autoUsers' is left to its default 'true' while we have the pwix:accounts-manager package too in the application
+// the autoUsers is a facility to make this package as transparent for possible for simple applications
+// as soon as we have a Accounts manager, we need amClass'es, and so must no leave this AccountsHub creates itself the 'users' collection
+Tracker.autorun(() => {
+    const autoUsers = AccountsHub.configure().autoUsers;
+    const haveManager = Object.keys( Package ).includes( 'pwix:accounts-manager' ) && Object.keys( Package['pwix:accounts-manager'] ).includes( 'AccountsManager' );
+    if( autoUsers && haveManager ){
+        logger.warning( ''
+            +'\'autoUsers\' is found left to its default configuration value (\'true\') while \'pwix:accounts-manager\' package is present: you should NOT keep both. '
+            +'If you need an Accounts manager, then leave it manage its accounts, and configure \'autoUsers\' to \'false\''
+        );
+    }
+});